Route each method to the good place.
